fix: rate 0 accepted

+++ backend/api/models/ratingModel.go
@@ -6,7 +6,7 @@ import (
 
 type Rating struct {
 	gorm.Model
-	RaterUsername string  `gorm:"NOT NULL"`
-	RatedUsername string  `gorm:"NOT NULL"`
-	Rate          float64 `gorm:"NOT NULL"`
+	RaterUsername string `gorm:"NOT NULL"`
+	RatedUsername string `gorm:"NOT NULL"`
+	Rate          int    `gorm:"NOT NULL"`
 }
